[CHORE] ApplicationPlugin 삭제 및 AndroidPlugin 통합
com.android.application 플러그인 이외 부분들이 겹치기 때문에 gh.application.gradle에서 해당 플러그인만 추가적으로 적용한 뒤, AndroidPlugin 적용
