+++ build-logic/src/main/java/gradle/gh.application.gradle.kts
@@ -1,5 +1,9 @@
-import plugin.configureApplication
+import plugin.configureAndroid
 import plugin.configureHilt
 
-configureApplication()
+plugins {
+    id("com.android.application")
+}
+
+configureAndroid()
 configureHilt()
